+++ commands/challonge/start-tournament.js
@@ -10,7 +10,7 @@ const {
     getTournamentIdByName,
     updateTournamentStatus,
     createTournamentChannels,
-} = require("../../testdatamanager");
+} = require("../../datamanager");
 const axios = require("axios");
 module.exports = {
     data: new SlashCommandBuilder()
