+++ commands/challonge/test-start-tournament.js
@@ -8,8 +8,8 @@ const {
     getNameFromChallongeId,
     getUserIdFromAutodartsId,
     getUserIdFromChallongeId,
-} = require("../../testdatamanager");
-const { getAllMatchesFromTournamentId } = require("../../testdatamanager");
+} = require("../../datamanager");
+const { getAllMatchesFromTournamentId } = require("../../datamanager");
 const {
     ChannelType,
     PermissionFlagsBits,
