+++ datamanager.js
